+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fix redundant public accessibility analysis false-positive for property wrappers.
 - Fix redundant public accessibility analysis false-positive for declarations referenced from a public `@inlinable` function.
 - Fix redundant public accessibility analysis false-positive for function parameter default values.
+- Fix redundant public accessibility analysis false-positive for inherited and default associated types.
 
 ## 2.17.1 (2023-12-04)